# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOCUMENTATION DE L'API DU PROJET MESHUTE",
        "description": "MESHUTE est un projet de gestion de ressources immobilières qui a pour principal but de rendre leur accès facile",
        "contact": {
            "email": "knk.towork@gmail.com"
        },
        "license": {
            "name": "Apache2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Url de la documentation swagger Meshute"
        }
    ],
    "paths": {
...
        "/logement": {
            "post": {
                "tags": [
                    "Logements"
                ],
                "summary": "Ajouter un nouveau logement",
                "description": "Fonctionnalité permettant au gérant d'ajouter un nouveau logement",
                "operationId": "addLogement",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_gerant",
                                    "type_logement",
                                    "description",
                                    "largeur",
                                    "longueur",
                                    "valeur",
                                    "ville",
                                    "quartier"
                                ],
                                "properties": {
                                    "id_gerant": {
                                        "type": "string"
                                    },
                                    "type_logement": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "largeur": {
                                        "type": "string"
                                    },
                                    "longueur": {
                                        "type": "string"
                                    },
                                    "valeur": {
                                        "type": "string"
                                    },
                                    "pays": {
                                        "type": "string"
                                    },
                                    "ville": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "type": "string"
                                    },
                                    "quartier": {
                                        "type": "string"
                                    },
                                    "rue": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logement ajouté"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/logement/{id_logement}": {
            "put": {
                "tags": [
                    "Logements"
                ],
                "summary": "Mise à jour",
                "description": "Mettre à jour les informations d'un logement spécifique",
                "operationId": "updateLogement",
                "parameters": [
                    {
                        "name": "id_logement",
                        "in": "path",
                        "description": "Identifiant du logement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_gerant",
                                    "type_logement",
                                    "description",
                                    "largeur",
                                    "longueur",
                                    "valeur",
                                    "ville",
                                    "quartier"
                                ],
                                "properties": {
                                    "id_gerant": {
                                        "type": "string"
                                    },
                                    "type_logement": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "largeur": {
                                        "type": "string"
                                    },
                                    "longueur": {
                                        "type": "string"
                                    },
                                    "valeur": {
                                        "type": "string"
                                    },
                                    "pays": {
                                        "type": "string"
                                    },
                                    "ville": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "type": "string"
                                    },
                                    "quartier": {
                                        "type": "string"
                                    },
                                    "rue": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logement ajouté"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/logement/{id_logement}/changer-visibilite": {
            "get": {
                "tags": [
                    "Logements"
                ],
                "summary": "Modifier la visibilité",
                "description": "Permet à l'utilisateur de définir si le logement doit être visible au grand public ou non",
                "operationId": "changeVisibility",
                "parameters": [
                    {
                        "name": "id_logement",
                        "in": "path",
                        "description": "Identifiant du logement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visibilité changée"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}